requirements.txt
python-dotenv>=1.0.0
google-genai>=1.0.0
google-api-core>=2.0.0
pandas>=2.0.0
requests>=2.28.0
lxml>=4.9.0
Pillow>=10.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
pytest>=8.0.0
